// clients/etx/src/core/transaction/types.ts
export const TRANSACTION_TYPE = {
  Legacy: 'legacy',
  Eip2930: 'eip2930',
  Eip1559: 'eip1559',
} as const